+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [0.3.0-dev0] - 19-02-18
+### Changed
+* todo
 
 ## [0.2.1-rc0] - 19-02-11
 ### Added